rename StubArchiveObjectMetadata to StubResolvedFormat
